requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
cffi==1.14.0
click==7.1.2
confuse==1.3.0
coverage==5.1
cryptography==2.9.2
csvmapper==0.7
Flask==1.1.2
flask-swagger-ui==3.25.0
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.3
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.6.8
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.1.0
